+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to the EngEmil STM32 Bootloader project will be documented i
 NB! Notes added here during development to keep track of changes.
 
 Added
+- **global/shared color** and brightness between solid, brightness, blinking and pulsation states.
 - New **powerdown animation sequence**.
 - New **powerup animation sequence**.
 - New **Render modes**; static, transitions, and continuous, to handle different types of modes and animations.